🎨 :: [#81] ProfileImageService를 ProfileService에 합병
