+++ Projects/Feature/Sources/Scene/Profile/ServeFile/ProfileImageServices.swift
@@ -1,58 +1,58 @@
-import Foundation
-import Moya
-
-struct ProfileImageResponse: Codable {
-    let file: Data
-}
-
-enum ProfileImageServices {
-    case submit(authorization: String, imageData: Data)
-    case update(authorization: String, imageData: Data)
-    case delete(authorization: String)
-}
-
-extension ProfileImageServices: TargetType {
-    var baseURL: URL {
-        return URL(string: "https://port-0-goms-backend-v2-duzu222alg58k27h.sel3.cloudtype.app/api/v2/account")!
-    }
-
-    var path: String {
-        switch self {
-        case .submit:
-            return "/image"
-        case .update:
-            return "/image"
-        case .delete:
-            return ""
-        }
-    }
-
-    var method: Moya.Method {
-        switch self {
-        case .submit:
-            return .post
-        case .update:
-            return .patch
-        case .delete:
-            return .delete
-        }
-    }
-
-    var task: Task {
-        switch self {
-        case let .submit(authorization, imageData),
-             let .update(authorization, imageData):
-            let formData = MultipartFormData(provider: .data(imageData), name: "File", fileName: "profile_image.jpg", mimeType: "image/jpeg")
-            return .uploadMultipart([formData])
-        case let .delete(authorization):
-            return .requestPlain
-        }
-    }
-
-    var headers: [String : String]? {
-        switch self {
-        default:
-            return ["Content-Type": "application/json"]
-        }
-    }
-}
+//import Foundation
+//import Moya
+//
+//struct ProfileImageResponse: Codable {
+//    let file: Data
+//}
+//
+//enum ProfileImageServices {
+//    case submit(authorization: String, imageData: Data)
+//    case update(authorization: String, imageData: Data)
+//    case delete(authorization: String)
+//}
+//
+//extension ProfileImageServices: TargetType {
+//    var baseURL: URL {
+//        return URL(string: "https://port-0-goms-backend-v2-duzu222alg58k27h.sel3.cloudtype.app/api/v2/account")!
+//    }
+//
+//    var path: String {
+//        switch self {
+//        case .submit:
+//            return "/image"
+//        case .update:
+//            return "/image"
+//        case .delete:
+//            return ""
+//        }
+//    }
+//
+//    var method: Moya.Method {
+//        switch self {
+//        case .submit:
+//            return .post
+//        case .update:
+//            return .patch
+//        case .delete:
+//            return .delete
+//        }
+//    }
+//
+//    var task: Task {
+//        switch self {
+//        case let .submit(authorization, imageData),
+//             let .update(authorization, imageData):
+//            let formData = MultipartFormData(provider: .data(imageData), name: "File", fileName: "profile_image.jpg", mimeType: "image/jpeg")
+//            return .uploadMultipart([formData])
+//        case let .delete(authorization):
+//            return .requestPlain
+//        }
+//    }
+//
+//    var headers: [String : String]? {
+//        switch self {
+//        default:
+//            return ["Content-Type": "application/json"]
+//        }
+//    }
+//}
